refactored index.js -- put renderBoard, renderUnits, handleMove, and updateUnits in their own files.  Added functions to updateSelectedUnit and setOffset values. Verified code is still working.
