+++ client/renderers/board.js
@@ -1,24 +1,61 @@
-export function renderBoard(gameboard) {
+import * as PIXI from 'pixi.js'
+import {
+  gameboard,
+  SCALE,
+  GameContainer,
+  selectedUnit,
+  updateSelectedUnit,
+  getOffset
+} from '../index'
+import {handleMove} from '../actions/move'
+
+//Making texture from image files
+const plainTile = PIXI.Texture.from('/images/terrain_plains.png')
+const mtTile = PIXI.Texture.from('/images/terrain_mountains.png')
+const tileTextures = {
+  plain: plainTile,
+  mountain: mtTile
+}
+
+//stores rendered sprites added to gameboard
+let tileSprites = []
+
+/*
+* * * * * * * * * * * * * * * * * * * * * * * * *
+ renderBoard external Dependencies:
+  gameboard
+  SCALE
+  offset()
+  PIXI
+  tileTextures
+  selectedUnit
+  handleMove
+  GameContainer
+  TileSprites
+* * * * * * * * * * * * * * * * * * * * * * * * *
+*/
+export function renderBoard() {
   //going through each row of board
-  for (let y = 0; y < gameboard.length; y++) {
+  for (let y = 0; y < gameboard.board.length; y++) {
     //offset for hex-style pattern
-    let offset = 0
-    if (y % 2 === 0) offset = SCALE / 2
+    let offset = getOffset(y)
 
     //going through each column of current row
-    for (let x = 0; x < gameboard[y].length; x++) {
+    for (let x = 0; x < gameboard.board[y].length; x++) {
       //make a new sprite
-      let tileSprite = new PIXI.Sprite(tileTextures[gameboard[y][x]])
+
+      //console.log(gameboard.board[y][x].value.name)
+
+      let tileSprite = new PIXI.Sprite(
+        tileTextures[gameboard.board[y][x].tile.name]
+      )
 
       //pass reference to tile into sprite
       // TYPES
       // 0 = clear
       // 1 = impassible
       //? Key-value data structure to id tile types
-      tileSprite.data = {
-        type: gameboard[y][x],
-        coordinates: {x, y}
-      }
+      tileSprite.data = gameboard.board[y][x]
 
       //scale & position
       tileSprite.width = SCALE
@@ -27,18 +64,21 @@ export function renderBoard(gameboard) {
       tileSprite.y = y * SCALE
 
       //rendering based on tile type
-      if (gameboard[y][x] === 0) tileSprite.tint = 0x008000
-      else if (gameboard[y][x] === 1) tileSprite.tint = 0xa52a2a
+      if (gameboard.board[y][x].tile.name === 'plain')
+        tileSprite.tint = 0x008000
+      else if (gameboard.board[y][x].tile.name === 'mountain')
+        tileSprite.tint = 0xa52a2a
 
       //setting event handlers
       tileSprite.interactive = true
 
       //onClick, call selectedUnit's move fn to clicked tile coord
-      tileSprite.on('click', e => {
-        if (selectedUnit.coordinates) {
-          // console.log('tile clicked, coordinates: ', tileSprite.data.coordinates)
-          selectedUnit.move(tileSprite.data.coordinates)
-          selectedUnit = {}
+      tileSprite.on('click', () => {
+        console.log('tile clicked, tile data: ', tileSprite.data)
+        if (selectedUnit.data.currentTile.coordinates) {
+          handleMove(selectedUnit, tileSprite.data)
+          //sets selectedUnit to an empty array
+          updateSelectedUnit({})
         }
       })
